% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proBAMr.R
\docType{package}
\name{proBAMr-package}
\alias{proBAMr}
\alias{proBAMr-package}
\title{proBAMr}
\description{
Maps peptide-spectrum-matches (PSMs) to a genome. The package builds SAM file
from shotgun proteomics data. The package also provides function
to prepare annotation from GTF file.
}
