% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMtab2SAM.R
\name{PSMtab2SAM}
\alias{PSMtab2SAM}
\title{Generate SAM files from PSMs.}
\usage{
PSMtab2SAM(passedPSM, exon_anno, proteinseq, procodingseq, snvproseq = NULL,
  snvprocoding = NULL, junpep = NULL, junpepcoding = NULL,
  jun_anno = NULL, outfile = "output.sam", show_progress = TRUE, ...)
}
\arguments{
\item{passedPSM}{a data.frame of PSMs (created by one of proBAMr's read functions)}

\item{exon_anno}{a data.frame of exon annotations}

\item{proteinseq}{a data.frame of protein ids and protein sequences}

\item{procodingseq}{a data.frame of coding sequences for each protein}

\item{snvproseq}{a data.frame of protein sequence changes due to simple nucleotide variations}

\item{snvprocoding}{a data.frame of coding changes due to simple nucleotide variations (SNPs and INDELs)}

\item{junpep}{a data.frame of peptide sequences coded for by novel splice junctions}

\item{junpepcoding}{a data.frame of coding sequences coded for by novel splice junctions}

\item{jun_anno}{a data.frame of genome annotation information for novel splice junctions}

\item{outfile}{the filepath of the output SAM file}

\item{show_progress}{if TRUE, long-running parts of the function will report progress information}

\item{...}{additional arguments}
}
\value{
invisible(NULL)
}
\description{
Generate SAM files from confident peptide-spectrum-matches (PSMs).
}
\examples{
load(system.file("extdata/GENCODE", "exon_anno.RData", package="proBAMr"))
load(system.file("extdata/GENCODE", "proseq.RData", package="proBAMr"))
load(system.file("extdata/GENCODE", "procodingseq.RData", package="proBAMr"))
passedPSM <- readPepXmlTab(system.file("extdata/psm/passedPSM.tab", package="proBAMr"))
PSMtab2SAM(passedPSM, exon, proteinseq,  procodingseq, outfile=paste0(tempdir(), '/test.sam'))
SAM <- read.table(paste0(tempdir(), '/test.sam'))

}
\author{
Xiaojing Wang
}
